Downloads the desk review template code

Downloads the desk review template code

Usage

copy_dr_template_code(
  output_path = Sys.getenv("DR_PATH"),
  country_name = Sys.getenv("DR_COUNTRY")
)

Arguments

output_path

str Where to download the desk review template code.

country_name

str Name of the country in the desk review.