Pull the main map data to build the map

This function primarily pulls the relevant information to build the map. Inherits areguments from generate_adhoc_map().

Usage

pull_map_data(
  raw.data,
  .vdpv,
  country,
  surv,
  virus_type,
  .start_date,
  .end_date
)

Arguments

raw.data

list global polio data output from get_all_polio_data()

.vdpv

logical Whether to include VDPV in maps.

country

str or list Name of the country or countries.

surv

str or list Surveillance type.

virus_type

str or list Virus type to map.

.start_date

date Start date.

.end_date

date End date.

Value

tibble Filtered dataset used to build the map.