sirfunctions: Key Functions to Analyze Global Polio Surveillance Data

The sirfunctions package contains key functions used by the Surveillance, Innovation, and Research (SIR) team within the Polio Eradication Branch (PEB) at the Centers for Disease Control and Prevention (CDC). It includes functions to download cleaned global polio data from the World Health Organization (WHO) Polio Information System (POLIS) and geographic files. In addition, the package contains functions to visualize important epidemiological trends and perform key performance indicators such as NPAFP rates, EV rates, and stool adequacy. Apart from getting data and calculating key indicators, there are several functions that also facilitate communication, visualize trends, and perform data quality checks.

See also

Author

Maintainer: Nishant Kishore ynm2@cdc.gov (ORCID)

Authors:

Mervin Keith Cuadera xrg9@cdc.gov (ORCID)

Nicholas Heaghney uic3@cdc.gov

Elizabeth Krow-Lucal yxn9@cdc.gov

Smita Chavan wsy2@cdc.gov